feat: partytown initialization to delayed
