+++ scripts/delayed.js
@@ -4,6 +4,14 @@ import loadVideoImpl from './assets/video-lib.js';
 import generateImageOverlayImpl from './assets/image-lib.js';
 import loadCookieConsent from './cookie-consent-lib.js';
 
+function initPartytown() {
+  window.partytown = {
+    lib: '/scripts/',
+    forward: ['dataLayer.push'],
+  };
+  import('./partytown.js');
+}
+
 // used by cookie-consent so far
 await fetchPlaceholders(`/${document.documentElement.lang}`);
 
@@ -11,6 +19,8 @@ loadCookieConsent();
 
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
+
+initPartytown();
 
 export default function loadVideo(videoURL, block) {
   loadVideoImpl(videoURL, block);
